+++ pypeit/telescopes.py
@@ -78,6 +78,13 @@ class LBTTelescopePar(TelescopePar):
                                                latitude=32.701308,
                                                elevation=3221.0)
 
+class KPNOTelescopePar(TelescopePar):
+    def __init__(self):
+        super(KPNOTelescopePar, self).__init__(name='KPNO',
+                                              longitude=111.616111,
+                                              latitude=31.9516666,
+                                              elevation=2098.)
+
 class MMTTelescopePar(TelescopePar):
     def __init__(self):
         super(MMTTelescopePar, self).__init__(name='MMT',